order, order-history

+++ Website_NoRulesShop/src/main/java/com/poly/website_norulesshop/Repository/OrderStatusRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderStatusRepository extends JpaRepository<OrderStatus, Long> {
+public interface OrderStatusRepository extends JpaRepository<OrderStatus, Integer> {
     // Các phương thức đặc thù cho bảng OrderStatus (nếu cần)
 }
